fix different warnings Delete TGeoid which is no longer used

+++ source/ProjectFramework/ProjectData/TWorkingPoints.cpp
@@ -144,7 +144,7 @@ bool TWorkingPoints::deletePoint(PointIterator iter)
 
 int TWorkingPoints::numberOfPoints() const{
 
-	return fWorkingPoints.size();
+	return (int)fWorkingPoints.size();
 }
